requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
python-multipart
passlib[bcrypt]
google-auth
requests